+++ crates/events/src/keyboard.rs
@@ -1,0 +1,9 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "kebab-case"))]
+#[non_exhaustive]
+/// A keyboard input.
+pub enum Keyboard {
+	/// Event representing an 'end of file' on stdin
+	Eof,
+}