+++ crates/lib/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next (YYYY-MM-DD)
 
 - Ditch MSRV policy. The `rust-version` indication will remain, for the minimum estimated Rust version for the code features used in the crate's own code, but dependencies may have already moved on. From now on, only latest stable is assumed and tested for. ([#510](https://github.com/watchexec/watchexec/pull/510))
+- Split off `watchexec-events` and `watchexec-signals` crates.
 
 ## v2.1.1 (2023-02-14)